Check if paths to sequence data refer to same file
If there are multiple paths to the sequence data, this change will pick the first path if all paths refer to the same file, as determined by a call to filecmp.cmp(), i.e. same creation time, same size, and same mode. This is useful for network shares that are mounted on different paths for different users, such as /home/user1/cifs/file1 vs /home/user2/cifs/file2.